LightningBeam Spell, Fix AI Chase, and Health and Souls per enemy tier

+++ Assets/Scripts/Characters/Player/PlayerController.cs
@@ -106,7 +106,7 @@ public class PlayerController : CharacterClass
 
     #endregion
 
-    public override void GetHit(int damageAmount,GameObject attacker, SpellBook spell)
+    public override void GetHit(float damageAmount,GameObject attacker, SpellBook spell)
     {
         base.GetHit(damageAmount, attacker, spell);
         if (health <= 0)
